src/bot/main.py
import discord
from discord.ext import commands
import os
from dotenv import load_dotenv

# load environment variables
load_dotenv()

# bot set up
intents = discord.Intents.default()
intents.message_content = True
bot = commands.Bot(command_prefix='!', intents=intents)

@bot.event
async def on_ready():
    print(f'{bot.user} has connected to Discord!')

# Run the bot
if __name__ == "__main__":
    bot.run(os.getenv('DISCORD_TOKEN'))

